commit 2020-05-15 삭제버튼 클릭시 한번더 물어보기

+++ src/main/webapp/WEB-INF/views/board/boardDetail.jsp
@@ -195,8 +195,11 @@
 		}
 		
 		function fnc_delete(){
-			var boardNo = getParameterByName("boardNo");
-			location.href="/deleteBoard.do?boardNo="+boardNo;
+			var result = confirm("이 글을 삭제하시겠습니까?");
+			if(result){
+				var boardNo = getParameterByName("boardNo");
+				location.href="/deleteBoard.do?boardNo="+boardNo;				
+			}
 		}
 		
 		/************************************************** file download */
